--- Content.Client/IdentityManagement/IdentitySystem.cs
@@ -1,7 +1,0 @@
-using Content.Shared.IdentityManagement;
-
-namespace Content.Client.IdentityManagement;
-
-public sealed class IdentitySystem : SharedIdentitySystem
-{
-}
