+++ Content.Server/Clothing/Systems/ChameleonClothingSystem.cs
@@ -1,9 +1,9 @@
 using System.Linq;
 using Content.Server.Emp;
-using Content.Server.IdentityManagement;
 using Content.Shared.Clothing.Components;
 using Content.Shared.Clothing.EntitySystems;
 using Content.Shared.Emp;
+using Content.Shared.IdentityManagement;
 using Content.Shared.IdentityManagement.Components;
 using Content.Shared.Inventory;
 using Content.Shared.Prototypes;
