+++ Content.Shared/StationRecords/StationRecordsComponent.cs
@@ -1,8 +1,6 @@
-using Content.Server.StationRecords.Systems;
+namespace Content.Shared.StationRecords;
 
-namespace Content.Server.StationRecords;
-
-[Access(typeof(StationRecordsSystem))]
+[Access(typeof(SharedStationRecordsSystem))]
 [RegisterComponent]
 public sealed partial class StationRecordsComponent : Component
 {
